jwt 유효기간 수정

+++ api/src/main/java/com/momato/util/JwtProperties.java
@@ -2,7 +2,7 @@ package com.momato.util;
 
 public class JwtProperties {
 	public static final String SECRET = "Secret";
-    public static final int EXPIRATION_TIME = 86400000*30; // 30 days
+    public static final long EXPIRATION_TIME = 86400000*30; // 30 days
     public static final String TOKEN_PREFIX = "Bearer ";
     public static final String HEADER_STRING = "Authorization";
 }
